split out expressions and statements into separate test files
